Remove testing files and move toString to lib
