+++ lib/toString.js
@@ -1,27 +1,27 @@
-const pluginUtils = require('./lib/pluginUtils')
+const pluginUtils = require('./pluginUtils')
 
 const LINES = {
-  r: '└──', //right
-  t: '├──', //T shape
-  v: '│  '  //vertical
+  r: '└──', // right
+  t: '├──', // T shape
+  v: '│  ' // vertical
 }
 
-function toString() {
-  let output_string = 'Fastify Instance toString'
+function toString () {
+  let outputString = 'Fastify Instance toString'
   /* Print hooks */
-  output_string += `\n\nTop Level Hooks`
+  outputString += `\n\nTop Level Hooks`
 
   Object.keys(this._hooks).forEach((lfm, i) => {
     const hooks = this._hooks[lfm]
-    output_string += `\n${
+    outputString += `\n${
       i < 3 ? LINES.t : LINES.r
-    } ${lfm} (${hooks.length} hooks): ${hooks.length > 0 ? hooks.map(h=>h.name) : '[]'}`
+    } ${lfm} (${hooks.length} hooks): ${hooks.length > 0 ? hooks.map(h => h.name) : '[]'}`
   })
   /* Ideally: iterate over routes and print plugins & middleware by scope */
-  output_string += `\n\nRoutes`
-  output_string += `\n${this.printRoutes()}`
+  outputString += `\n\nRoutes`
+  outputString += `\n${this.printRoutes()}`
 
-  output_string += `\nTop Level Middleware`
+  outputString += `\nTop Level Middleware`
   const middlewares = {
     '/': []
   }
@@ -33,21 +33,21 @@ function toString() {
     } else middlewares['/'].push(mdw[0].name)
   })
   Object.keys(middlewares).forEach((key, i, keys) => {
-    output_string += `\n${
+    outputString += `\n${
       i < keys.length - 1 ? LINES.t : LINES.r
     } ${key}: [${middlewares[key].join(', ')}]`
   })
 
-  output_string += `\n\nPlugins`
+  outputString += `\n\nPlugins`
   const plugins = this[pluginUtils.registeredPlugins]
   plugins.forEach((plugin, i) => {
-    output_string += `\n${
+    outputString += `\n${
       i < plugins.length - 1 ? LINES.t : LINES.r
     } ${plugin.name.length === 0 ? 'anonymous' + i : plugin.name}`
   })
   // console.log(this)
-  console.log(output_string)
-  return output_string
+  console.log(outputString)
+  return outputString
 }
 
 module.exports = toString
